add constants, some fixes and validations

+++ internal/mutants/model.go
@@ -3,5 +3,5 @@ package mutants
 type Stats struct {
 	MutantDNACount int64   `json:"count_mutant_dna"`
 	HumanDNACount  int64   `json:"count_human_dna"`
-	Ratio          float64 `json:"ratio"`
+	Ratio          float32 `json:"ratio"`
 }
